+++ recipes/default.rb
@@ -6,6 +6,7 @@
 #
 
 include_recipe 'lemp-server::system'           # system basic setup (repos, tweaks)
+include_recipe 'lemp-server::ruby'             # ruby
 include_recipe 'lemp-server::web-db'           # Web: MySQL
 include_recipe 'lemp-server::web-nginx'        # Web: Nginx server
 include_recipe 'lemp-server::web-php'          # Web: PHP, PHP-FPM, composer, phpMyAdmin